@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>    
-    <TargetFramework>net4.7</TargetFramework>    
+    <TargetFramework>net4.8</TargetFramework>    
     <IsPackable>false</IsPackable>
     <Platforms>AnyCPU;x64;x86</Platforms>
   </PropertyGroup>
