@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <UserSecretsId>3ac0dbeb-4d8a-479e-b111-5efb8f1d0e6d</UserSecretsId>
     <DockerDefaultTargetOS>Linux</DockerDefaultTargetOS>
     <DockerfileContext>..\..</DockerfileContext>
@@ -12,7 +12,7 @@
     <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="$(AutoMapperDILib)" />
     <PackageReference Include="Macross.Json.Extensions" Version="$(MacrossJson)" />
 
-    <PackageReference Include="Microsoft.Extensions.PlatformAbstractions" Version="1.1.0" />
+    <PackageReference Include="Microsoft.VisualStudio.Azure.Containers.Tools.Targets" Version="$(AzureContainer)" />
 
     <PackageReference Include="Swashbuckle.AspNetCore" Version="$(Swashbuckle)" />
     <PackageReference Include="Swashbuckle.AspNetCore.SwaggerGen" Version="$(Swashbuckle)" />
