@@ -1,6 +1,6 @@
-﻿#region Copyright (C) 2017-2021 Yaroslav Tatarenko
+﻿#region Copyright (C) 2017-2022 Yaroslav Tatarenko
 
-// Copyright (C) 2017-2021 Yaroslav Tatarenko
+// Copyright (C) 2017-2022 Yaroslav Tatarenko
 // This product uses MediaInfo library, Copyright (c) 2002-2021 MediaArea.net SARL. 
 // https://mediaarea.net
 
@@ -9,31 +9,30 @@
 using System.Runtime.Serialization;
 using System.Text.Json.Serialization;
 
-namespace ApiSample.Models
+namespace ApiSample.Models;
+
+/// <summary>
+/// Describes bitrate possible modes
+/// </summary>
+[DataContract]
+[JsonConverter(typeof(JsonStringEnumMemberConverter))]
+public enum BitrateMode : byte
 {
     /// <summary>
-    /// Describes bitrate possible modes
+    /// Constant quality mode
     /// </summary>
-    [DataContract]
-    [JsonConverter(typeof(JsonStringEnumMemberConverter))]
-    public enum BitrateMode : byte
-    {
-        /// <summary>
-        /// Constant quality mode
-        /// </summary>
-        [EnumMember(Value = "constantQuality")]
-        Cq,
+    [EnumMember(Value = "constantQuality")]
+    Cq,
 
-        /// <summary>
-        /// Constant bitrate mode
-        /// </summary>
-        [EnumMember(Value = "constantBitrate")]
-        Cbr,
+    /// <summary>
+    /// Constant bitrate mode
+    /// </summary>
+    [EnumMember(Value = "constantBitrate")]
+    Cbr,
 
-        /// <summary>
-        /// Variable bitrate mode
-        /// </summary>
-        [EnumMember(Value = "variableBitrate")]
-        Vbr
-    }
+    /// <summary>
+    /// Variable bitrate mode
+    /// </summary>
+    [EnumMember(Value = "variableBitrate")]
+    Vbr
 }
